refactor: move action authorization
