+++ packages/agent/test/__factories__/authorization/authorization.ts
@@ -6,9 +6,6 @@ import forestAdminClientFactory from '../forest-admin-client';
 export class AuthorizationsFactory extends Factory<AuthorizationService> {
   mockAllMethods() {
     return this.afterBuild(Authorizations => {
-      Authorizations.assertCanApproveCustomAction = jest.fn();
-      Authorizations.assertCanRequestCustomActionParameters = jest.fn();
-      Authorizations.assertCanTriggerCustomAction = jest.fn();
       Authorizations.assertCanBrowse = jest.fn();
       Authorizations.assertCanRead = jest.fn();
       Authorizations.assertCanAdd = jest.fn();
@@ -18,7 +15,6 @@ export class AuthorizationsFactory extends Factory<AuthorizationService> {
       Authorizations.getScope = jest.fn();
       Authorizations.assertCanExecuteChart = jest.fn();
       Authorizations.invalidateScopeCache = jest.fn();
-      Authorizations.verifySignedActionParameters = jest.fn();
     });
   }
 }
